Move functions to TLuaInterpreterMudletObjects.cpp
